Fixed some crashing issues. Only firing if PlayerConnectionController != null in player (not triggering with AvtomatV1)
