+++ Game/src/java/Controller/ConnectionController.java
@@ -83,7 +83,7 @@ public class ConnectionController implements ConnectionInterface {
             case 156 -> {
                 System.out.println("Modelo recibido");
                 playerConnController.setPlayerModel(packet.getObject().toString(), packet.getSourceID());
-                this.notifyPlayerJoin();
+
             }
             case 300 -> {
                 GameRules.numPlayers++;
@@ -115,41 +115,6 @@ public class ConnectionController implements ConnectionInterface {
 
     @Override
     public void onLookupUpdate(ArrayList<String> macs) {
-
-    }
-
-
-    /**
-     * Sends packet to all ConnectedPCs to update the amount of total players and updates this pc's total players.
-     */
-    private void notifyPlayerJoin() {
-        //Gets all Mac Addresses (not LocalMac)
-        for (String mac : comController.getConnectedMacs()) {
-            //Checks if Mac is PC
-            if (comController.getConnectedDeviceType(mac) == CommunicationController.PC) {
-                comController.sendMessage(comController.createPacket(mac, 300, null));
-
-            }
-        }
-        GameRules.numPlayers++;
-        System.out.println("Added Local Player " + GameRules.numPlayers);
-
-    }
-
-    /**
-     * Sends packet to all ConnectedPCs to update the amount of total players and updates this pc's total players.
-     */
-    public void notifyPlayerDeath() {
-        //Gets all Mac Addresses (not LocalMac)
-        for (String mac : comController.getConnectedMacs()) {
-            //Checks if Mac is PC
-            if (comController.getConnectedDeviceType(mac) == CommunicationController.PC) {
-                comController.sendMessage(comController.createPacket(mac, 301, null));
-
-            }
-        }
-        GameRules.numPlayers--;
-        System.out.println("Added Local Player " + GameRules.numPlayers);
 
     }
 }
